add multi uom test cases
