+++ regression10.xml
@@ -14,6 +14,17 @@
             <class name="com.cutanddry.qa.tests.price.ValidateThePoundSpotPricingTest"/>
             <class name="com.cutanddry.qa.tests.price.ValidateTheSpotPricingNotFromPricingEngineTest"/>
             <class name="com.cutanddry.qa.tests.price.ValidateTheSplitWeightTest"/>
+
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyEditedQuantitiesAfterEditingMultipleUOMInPDPTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyNavigationPDPWhenSelectingMultipleUnitsOptionTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyTheDeleteItemFromMultipleUOMInReviewOrderTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyTheEditQuantityOfMultipleUOMInOGAndSubmissionTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyTheEditQuantityOfMultipleUOMInOrderGuideTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyTheEditQuantityOfMultipleUOMInPDPAndSubmissionTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyTheEditQuantityOfMultipleUOMInPDPTest"/>
+            <class name="com.cutanddry.qa.tests.multi_uom.VerifyTheEditQuantityOfMultipleUOMInReviewOrderTest"/>
+
+
         </classes>
     </test>
 </suite>
