+++ src/main/java/com/cutanddry/qa/data/testdata/CatalogData.java
@@ -1,0 +1,11 @@
+package com.cutanddry.qa.data.testdata;
+
+public class CatalogData {
+    public static final String CUSTOMER_ID = "16579";
+    public static final String ITEM_CODE = "01409";
+    public static final String ITEM_NAME = "Carrot - Baby Peeled - 1 LB";
+    public static final String ADDED_QUANTITY = "Added: 1 CS | 1 Pkg";
+    public static final String MULTI_UOM_1 = "1";
+    public static final String MULTI_UOM_2 = "2";
+    public static final String UOM_DROPDOWN_OPTION = "Multiple Units";
+}
