+++ src/test/java/com/cutanddry/qa/tests/multi_uom/VerifyTheEditQuantityOfMultipleUOMInOGAndSubmissionTest.java
@@ -1,7 +1,8 @@
-package com.cutanddry.qa.tests.Multi_UOM;
+package com.cutanddry.qa.tests.multi_uom;
 
 import com.cutanddry.qa.base.TestBase;
 import com.cutanddry.qa.data.models.User;
+import com.cutanddry.qa.data.testdata.CatalogData;
 import com.cutanddry.qa.functions.Catalog;
 import com.cutanddry.qa.functions.Customer;
 import com.cutanddry.qa.functions.Dashboard;
@@ -16,11 +17,11 @@ import org.testng.asserts.SoftAssert;
 public class VerifyTheEditQuantityOfMultipleUOMInOGAndSubmissionTest extends TestBase {
     SoftAssert softAssert;
     static User user;
-    static String customerId = "16579";
-    String searchItemCode = "01700";
+    static String customerId = CatalogData.CUSTOMER_ID;
+    String searchItemCode = CatalogData.ITEM_CODE;
     static double itemOGPriceUOM1 ,itemOGPriceUOM2,totalOGItemPrice1;
-    String uom1 = "1";
-    String uom2 = "2";
+    String uom1 = CatalogData.MULTI_UOM_1;
+    String uom2 = CatalogData.MULTI_UOM_2;
     static String orderId;
 
     @BeforeMethod
